Attendance Pulling

+++ percent_data.py
@@ -1,3 +1,3 @@
 def for_(data, per):
     for i in data:
-        print(i[0], i[1] * (100 / per) - i[2])
+        print(i[0], int(i[1] * (100 / per) - i[2]))
